+++ CHANGELOG.md
@@ -75,6 +75,7 @@ All notable changes to this project will be documented in this file.
 - In WPF, make sure the axes are initalized when the Model is set before the PlotView has been loaded (#1303)
 - Fixed MinimumSegmentLength not working for LineSeries (#1044)
 - Fixed rendering issues with MagnitudeAxisFullPlotArea (#1364)
+- OxyPlot.Core.Drawing PngExporter Export to Stream background (#1382)
 
 ## [1.0.0] - 2016-09-11
 ### Added